AARGH. Godamn file decriptor leak which has been driving me CRAZY for a WEEK! Got the BASTARD. DIE!
(what complicated matters was that it only leaked on when it *couldn't* load
the image).

Please note. there is also a filedes leak in the tiff loader, also when it
fails to load. I haven't been able to fix this yet, because libtiff sucks
ass and has totally encapsulated the file descriptor. Lemme work on it
tomorrow...


SVN revision: 2267