An XCF loader. Currently it can handle layers, layer offsets, layer opacity, layer masks, and merging layers in the default mode (simply "looking" through all the layers). The other layer modes are missing right now (I hardly ever use anything other than "Normal" anyway, but that's just me of course).
If you load an image as /path/to/file:<layer number> only the
specified layer is displayed, at the defined offset in the image.
Try this with imlib2_view since it ignores file loading errors.
Passing parameters to the loader will have to be handled better
(or did I miss someting?) but the loader can basically return
layers selectively.


SVN revision: 2847