Urm. oops. hehe :)
Reversed that change, as it caused segfaults in free().

I can't grok this code, so I'll leave it to raster.

Basically, if my code calls imlib_free_image_and_decache(), it doesn't free
the image, and it leaks.

The fix I applied made sure _Imlib_ConsumeImage was called, but this caused
a segfault.

*baffle


SVN revision: 3239